route-optimization-engine/requirements.txt
# Route Optimization Engine - Python Dependencies

# Optimization
ortools==9.9.3963
numpy==1.26.4
scipy==1.12.0
pandas==2.2.1

# Geospatial
geopy==2.4.1
shapely==2.0.3

# Database Connectors
pymongo==4.6.2
snowflake-connector-python==3.7.1

# Salesforce Integration
simple-salesforce==1.12.6

# Data Generation
faker==24.1.0

# Visualization
matplotlib==3.8.3
folium==0.16.0

# API / HTTP
requests==2.31.0
flask==3.0.2

# Testing
pytest==8.1.1
pytest-cov==5.0.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
tqdm==4.66.2
